requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
pandas_datareader
requests
streamlit
python-dotenv
transformers
torch
